% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grass_plot}
\alias{grass_plot}
\title{Plot a grass object}
\usage{
grass_plot(x, ...)
}
\arguments{
\item{x}{A grass object.}

\item{...}{Passed through to the class-specific plot method.}
}
\value{
A ggplot object.
}
\description{
Dispatches to \code{plot.grass_result} or \code{plot.grass_reference}.
}
